+++ mayan/apps/cabinets/tests/test_wizard_steps.py
@@ -1,7 +1,7 @@
 from mayan.apps.documents.models import Document
 from mayan.apps.documents.permissions import permission_document_create
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
-from mayan.apps.sources.tests.mixins.web_form_source_mixins import WebFormSourceTestMixin
+from mayan.apps.sources.tests.mixins.web_form_source_mixins import WebFormSourceBackendTestMixin
 from mayan.apps.sources.wizards import DocumentCreateWizardStep
 
 from ..wizard_steps import DocumentCreateWizardStepCabinets
@@ -11,7 +11,7 @@ from .mixins import CabinetDocumentUploadWizardStepTestMixin, CabinetTestMixin
 
 class CabinetDocumentUploadTestCase(
     CabinetTestMixin, CabinetDocumentUploadWizardStepTestMixin,
-    WebFormSourceTestMixin, GenericDocumentViewTestCase
+    WebFormSourceBackendTestMixin, GenericDocumentViewTestCase
 ):
     auto_upload_test_document = False