+++ mayan/apps/tags/tests/test_wizard_steps.py
@@ -1,14 +1,14 @@
 from mayan.apps.documents.models import Document
 from mayan.apps.documents.permissions import permission_document_create
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
-from mayan.apps.sources.tests.mixins.web_form_source_mixins import WebFormSourceTestMixin
+from mayan.apps.sources.tests.mixins.web_form_source_mixins import WebFormSourceBackendTestMixin
 
 from .mixins import TagTestMixin, TaggedDocumentUploadWizardStepViewTestMixin
 
 
 class TaggedDocumentUploadViewTestCase(
     TaggedDocumentUploadWizardStepViewTestMixin, TagTestMixin,
-    WebFormSourceTestMixin, GenericDocumentViewTestCase
+    WebFormSourceBackendTestMixin, GenericDocumentViewTestCase
 ):
     auto_upload_test_document = False